+++ packages/sitecore-jss-angular/src/components/placeholder.component.ts
@@ -22,7 +22,7 @@ import {
   ViewChild,
   ViewContainerRef,
 } from '@angular/core';
-import { Data, Router, UrlTree } from '@angular/router';
+import { Data, RedirectCommand, Router, UrlTree } from '@angular/router';
 import { ComponentRendering, HtmlElementRendering } from '@sitecore-jss/sitecore-jss/layout';
 import { Observable } from 'rxjs';
 import { takeWhile } from 'rxjs/operators';
@@ -251,7 +251,10 @@ export class PlaceholderComponent implements OnInit, OnChanges, DoCheck, OnDestr
         this.isLoading = false;
         if (e instanceof JssCanActivateRedirectError) {
           const redirectValue = e.redirectValue;
-          if (redirectValue instanceof UrlTree) {
+
+          if (redirectValue instanceof RedirectCommand) {
+            this.router.navigateByUrl(redirectValue.redirectTo);
+          } else if (redirectValue instanceof UrlTree) {
             this.router.navigateByUrl(redirectValue);
           } else if (typeof redirectValue === 'string') {
             this.router.navigate([redirectValue]);